+++ README.md
@@ -1,12 +1,30 @@
-# MyGit
+# G
 
 ## About
 
-Playing around writing a toy git
-
-
-## Init
+An experimental `git` client written in Go.
 
 ```
+Usage:
+   [command]
 
+Available Commands:
+  add         
+  branch      
+  commit      
+  completion  Generate the autocompletion script for the specified shell
+  help        Help about any command
+  init        
+  log         
+  ls-files    
+  restore     
+  status      
+  switch      
+
+Flags:
+      --git-directory string   --git-directory (default ".git")
+  -h, --help                   help for this command
+      --path string            --path (default ".")
+
+Use " [command] --help" for more information about a command.
 ```
