+++ cmd/commit.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"encoding/hex"
 	"fmt"
-	"github.com/richardjennings/mygit/pkg/mygit/git"
+	"github.com/richardjennings/mygit/pkg/git"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
