+++ pkg/git/add_test.go
