+++ pkg/git/switch_test.go
