+++ pkg/index/index_darwin.go
